fix for dl build
reason for issue is incorrect location of core_prov structure.
fix is move core_prov declaration into common source file,
where it is available from any build configuration
