+++ include/fi.h
@@ -119,6 +119,7 @@ struct fi_filter {
 };
 
 extern struct fi_filter prov_log_filter;
+extern struct fi_provider core_prov;
 
 void fi_create_filter(struct fi_filter *filter, const char *env_name);
 void fi_free_filter(struct fi_filter *filter);
